+++ examples/test-service-node-host.js
@@ -8,19 +8,21 @@ class ExampleService extends Dataparty.IService {
   constructor(opts){
     super(opts)
 
-    this.addMiddleware(Dataparty.middleware_paths.pre.decrypt)
-    this.addMiddleware(Dataparty.middleware_paths.pre.validate)
+    let builder = new Dataparty.ServiceBuilder(this)
 
-    this.addMiddleware(Dataparty.middleware_paths.post.validate)
-    this.addMiddleware(Dataparty.middleware_paths.post.encrypt)
+    builder.addMiddleware(Dataparty.middleware_paths.pre.decrypt)
+    builder.addMiddleware(Dataparty.middleware_paths.pre.validate)
 
-    this.addEndpoint(Dataparty.endpoint_paths.echo)
-    this.addEndpoint(Dataparty.endpoint_paths.secureecho)
-    this.addEndpoint(Dataparty.endpoint_paths.identity)
-    this.addEndpoint(Dataparty.endpoint_paths.version)
+    builder.addMiddleware(Dataparty.middleware_paths.post.validate)
+    builder.addMiddleware(Dataparty.middleware_paths.post.encrypt)
 
-    this.addSchema(Path.join(__dirname, './party/schema/basic_types.js'))
-    this.addTopic(Path.join(__dirname, './party/topics/time-topic.js'))
+    builder.addEndpoint(Dataparty.endpoint_paths.echo)
+    builder.addEndpoint(Dataparty.endpoint_paths.secureecho)
+    builder.addEndpoint(Dataparty.endpoint_paths.identity)
+    builder.addEndpoint(Dataparty.endpoint_paths.version)
+
+    builder.addSchema(Path.join(__dirname, './party/schema/basic_types.js'))
+    builder.addTopic(Path.join(__dirname, './party/topics/time-topic.js'))
   }
 
 }
@@ -33,7 +35,8 @@ async function main(){
 
   
   const service = new ExampleService({ name: '@dataparty/example', version: '0.0.1' })
-  const build = await service.compile(Path.join(__dirname,'/dataparty'), true)
+  const builder = new Dataparty.ServiceBuilder(service)
+  const build = await builder.compile(Path.join(__dirname,'/dataparty'), true)
   
   const serviceName = build.package.name
   const basePath = '/data/dataparty/'
